changed file structure to separate chat and freelancer controller
