+++ routes/freelancerRoutes.js
@@ -8,10 +8,6 @@ router.get("/freelancers/projects", freelancerController.getFreelancerProjects);
 // Route to get all projects
 router.get("/projects", freelancerController.getAllProjects);
 
-//route to create chats
-router.post("/create-chat", freelancerController.createChat);
-
-//route to delete a chat
-router.delete("delete-chat", freelancerController.deleteChat);
+router.get("/freelancers", freelancerController.getFreelancers);
 
 module.exports = router;
